decompose part one
