+++ src/components/const.js
@@ -23,4 +23,4 @@ export const MONTHS = [
 ];
 
 export const COLORS = [`black`, `yellow`, `blue`, `green`, `pink`];
-export const DAYS = ["mo", "tu", "we", "th", "fr", "sa", "su"];
+export const DAYS = [`mo`, `tu`, `we`, `th`, `fr`, `sa`, `su`];
